requirements.txt
leidenalg
igraph
pillow
numpy